fix: resolve lint failures and rename utils.py to display.py (#149)
